+++ etc/blocks/sum-numbers/model/src/index.ts
@@ -2,11 +2,7 @@ import type {
   InferHrefType,
   InferOutputsType } from '@platforma-sdk/model';
 import {
-  Args,
   BlockModel,
-  getJsonField,
-  isEmpty,
-  not,
   PlRef,
 } from '@platforma-sdk/model';
 import { z } from 'zod';