+++ tools/tengo-builder/eslint.config.mjs
@@ -4,4 +4,9 @@ import { node } from '@milaboratories/eslint-config';
 export default [
   { ignores: ['src/**/*.test.ts', 'test-assets/**'] },
   ...node,
+  {
+    rules: {
+      '@typescript-eslint/require-await': 'off',
+    },
+  },
 ];