+++ etc/blocks/pool-explorer/block/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @milaboratories/milaboratories.pool-explorer
 
+## 1.0.208
+
+### Patch Changes
+
+- Updated dependencies [d60b0fe]
+  - @platforma-sdk/model@1.37.18
+  - @milaboratories/milaboratories.pool-explorer.workflow@1.1.38
+  - @milaboratories/milaboratories.pool-explorer.ui@1.1.141
+  - @milaboratories/milaboratories.pool-explorer.model@1.0.60
+
+## 1.0.207
+
+### Patch Changes
+
+- @milaboratories/milaboratories.pool-explorer.ui@1.1.140
+
+## 1.0.206
+
+### Patch Changes
+
+- @milaboratories/milaboratories.pool-explorer.workflow@1.1.38
+
+## 1.0.205
+
+### Patch Changes
+
+- @milaboratories/milaboratories.pool-explorer.ui@1.1.139
+
 ## 1.0.204
 
 ### Patch Changes