+++ lib/node/pl-deployments/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @milaboratories/pl-local
 
+## 2.4.3
+
+### Patch Changes
+
+- e6d4998: Backend version upgraded to 1.35.2
+
+## 2.4.2
+
+### Patch Changes
+
+- Updated dependencies [d60b0fe]
+  - @milaboratories/pl-config@1.6.1
+
 ## 2.4.1
 
 ### Patch Changes