+++ lib/node/pl-drivers/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @milaboratories/pl-drivers
 
+## 1.5.78
+
+### Patch Changes
+
+- d60b0fe: Chore: fix linter errors
+- Updated dependencies [d60b0fe]
+  - @milaboratories/pl-tree@1.6.12
+
+## 1.5.77
+
+### Patch Changes
+
+- Updated dependencies [e210414]
+  - @milaboratories/helpers@1.6.17
+
 ## 1.5.76
 
 ### Patch Changes